+++ portal-impl/src/com/liferay/portal/events/SampleAppStartupAction.java
@@ -70,8 +70,8 @@ public class SampleAppStartupAction extends SimpleAction {
 		String firstName = "Paul";
 		String middleName = StringPool.BLANK;
 		String lastName = "Smith";
-		int prefixId = 0;
-		int suffixId = 0;
+		long prefixId = 0;
+		long suffixId = 0;
 		boolean male = true;
 		int birthdayMonth = Calendar.JANUARY;
 		int birthdayDay = 1;
